+++ CHANGELOG.md
@@ -99,3 +99,5 @@
 ## 3/3/25
 - Added RPM stats to the trip summary screen
 - Added proper efficiency score calculation based on the speed,fuel and rpm
+- Added the score to the trip card screens
+- Added rpm info on the cards too